pharmacy-rag-assistant/requirements.txt
openai>=1.0.0
langchain>=0.1.0
langchain-openai>=0.0.2
psycopg2-binary>=2.9.0
pgvector>=0.2.0
sqlalchemy>=2.0.0
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.14.0
matplotlib>=3.7.0
seaborn>=0.12.0
streamlit>=1.28.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pytest>=7.4.0
pytest-asyncio>=0.21.0